+++ tests/testthat/test-read-ac-saf-uv-hdf5.R
@@ -172,7 +172,6 @@ test_that("reads two different grid data files", {
 
 })
 
-
 test_that("errors are triggered", {
 
   one.file.name <-